Tried to fix MakeFile
I have an issue with my makefile where it will not work on Windows (because the search command is not available). However, if windows users have make installed, I want them to be able to run the make script. There is even an analog command for this: dir /s/b *.java > sources. However, for some reason, when I run the os check and it gets to actually running the dir command, it will use the sh version of it. I want to use the windows version, and I don't have enough time right now to research how to do this. So, to stack overflow I go asking another question.