Worked on Documentation
Worked on documentation. However, there is a [bug](https://github.com/assertj/assertj-core/issues/1403) in the javadoc command for openjdk 11, and the laptop I'm working on doesn't allow for a 64 bit download of the jdk (which is what I need). I <b>never</b> thought that I'd encounter an issue like this.