Fixed Doc tag in Makefile
So the concensus on how to fix the documentation issue was to move from the Java 8 docs to the Java 15 docs (such a jump I know). The reason is due to the fact that the url changed a bit. In the 15 version, they added a /en/java/ right behind the .com, and also allowed for java modules to be included by including a /java.base/ url within that as well. And while this change can be seen with the javadoc 11 tool, i haven't been able to find the documentation for the javadoc tool at all, save for the java 7 documentation. So I had to reverse engineer what the hell was going on. Essensially, Java 11 broke the javadoc tool and didn't tell us HOW it broke it; the current makefile works for Java 11 and up.