Coding Style Fix (#9)
* Documented com.coswald.jtalker Package

Started working on the Documentation issue by commenting com.coswald.jtalker. This is a far cry to getting the actual thing fully documented; however, this is a good start to making sure all commits from this point are signed and legit as well.

* Worked on Documentation

Worked on documentation. However, there is a [bug](https://github.com/assertj/assertj-core/issues/1403) in the javadoc command for openjdk 11, and the laptop I'm working on doesn't allow for a 64 bit download of the jdk (which is what I need). I <b>never</b> thought that I'd encounter an issue like this.

* Fixed Doc tag in Makefile

So the concensus on how to fix the documentation issue was to move from the Java 8 docs to the Java 15 docs (such a jump I know). The reason is due to the fact that the url changed a bit. In the 15 version, they added a /en/java/ right behind the .com, and also allowed for java modules to be included by including a /java.base/ url within that as well. And while this change can be seen with the javadoc 11 tool, i haven't been able to find the documentation for the javadoc tool at all, save for the java 7 documentation. So I had to reverse engineer what the hell was going on. Essensially, Java 11 broke the javadoc tool and didn't tell us HOW it broke it; the current makefile works for Java 11 and up.

* Finished Documentation

Finished the documentation of TCPClientInstance and any others I missed. It was a great journey working on documentation, and I will never forget why I do this while I'm going ever again. I had to remember why I made certain decisions on certain things, and I didn't self-document well. All's well that ends well I guess.

* Added Coding Style

Added a coding style, which follows everything I currently or will do. I think there may be one specific thing I need to look at (Law of Demeter in regards to the ServerOutputStream), but other than that I think I'm good.

* Fixed Readme

Fixed some things within the README that just make sense. One thing that we fixed was the column length, and we also made sure the file was readable. Mostly, we gave credit where credit was due for the coding style.