Added ColorChooserPanel
Added a color choosing panel. This panel is seen at the bottem left of the JTalker application, and I thought it would be the hardest to design. Luckily, we "hacked" the JColorChooser to allow us to just reposition the elements of said panel. This saved us a significant amount of design choices and limitations; it will also allow us to work on the rest of the UI without worrying about this one. Overall, I like how it turned out. The only different between the design image and the actual panel we created is the fact that the preview color is just a banner on top of the color choosing panel. I think it will look really cool, because it will be right below the canvas.