Added CanvasPanel
Added the Canvas Panel, which allows for drawing with different colors. However, so far this will only work for the client. When we add the UDP server, we will need to update this class as to allow for multiple UDPClientInstances to change it. How we need to do this is going to be a point of research.