Added the ColoredTextPane
Added the ColoredTextPane, which is a JTextPane which allows colors defined by the ANSI color code standard. This is awesome, as it will allow a JTalker application to receive text from the server and set the colors based on codes received by the server (at little to no change from the server). What is even better is the fact that this is extendable by the ANSI color code standard: we currently only have background and foreground colors implemented. However, we could have bold or italic, or underline, or concealed text. It really is exciting :)