Fixed TextPipe bug
I didn't know there was a bug within the text pipe implementation, but I found it and fixed it. Basically, every time you were to add another text updater to the text entry panel, it woul donly work for one of them. But then I realized the issue: the text updater was clearing the text before the others can use it. I could've been lazy and removed that part of the action listener altogether and then only removed it when the enter button was pressed; in the end, I settled for allowing the button to remove the text as its own action (and ensuring this action happened last).