Implemented GUI (#15)
* Added ColorChooserPanel

Added a color choosing panel. This panel is seen at the bottem left of the JTalker application, and I thought it would be the hardest to design. Luckily, we "hacked" the JColorChooser to allow us to just reposition the elements of said panel. This saved us a significant amount of design choices and limitations; it will also allow us to work on the rest of the UI without worrying about this one. Overall, I like how it turned out. The only different between the design image and the actual panel we created is the fact that the preview color is just a banner on top of the color choosing panel. I think it will look really cool, because it will be right below the canvas.

* Added GUI package-info.java

Ooops. Forgot this class. I also changed the version from 0.2.0 to 0.1.5, because we aren't actually done with 0.2.0.

* Added GUI package-info.java

Ooops. Forgot this class. I also changed the version from 0.2.0 to 0.1.5, because we aren't actually done with 0.2.0.

* Fixed Some Codacy Issues

I
've been needing to do this. I just needed to fix those few Codacy issues that I can fix.

* Fixed More Codacy Issues

I've been needing to do this. I just needed to fix those few Codacy issues that I can fix.

* Redesigned ColorChooserPanel.ColorUpdater

Redesigned the ColorChooserPanel.ColorUpdater to make it so that any color can be changed by the color chooser, mainly using a setColor(Color) method.

* Added CanvasPanel

Added the Canvas Panel, which allows for drawing with different colors. However, so far this will only work for the client. When we add the UDP server, we will need to update this class as to allow for multiple UDPClientInstances to change it. How we need to do this is going to be a point of research.

* Added the ColoredTextPane

Added the ColoredTextPane, which is a JTextPane which allows colors defined by the ANSI color code standard. This is awesome, as it will allow a JTalker application to receive text from the server and set the colors based on codes received by the server (at little to no change from the server). What is even better is the fact that this is extendable by the ANSI color code standard: we currently only have background and foreground colors implemented. However, we could have bold or italic, or underline, or concealed text. It really is exciting :)

* Fixed Codacy Issues

Fixed Codacy issues with last commit. There is still the issue of the getANSIColor method being too long; we will fix this at a later date.

* Added TextEntryPanel

Added the text entry panel, which allows us to send text to multiple places via 'pipes' that are described by the TextPipe interface.

* Fixed Drawing Bug

I found out that if you just clicked the CanvasPanel, it would not draw that point until a new curve was drawn. So I fixed that

* Added escape sequence testing to ANSIColorConstants

Added escape sequence testing to ANSI color constants class.

* Added Printing Functionality

Added printing functionality to the ColoredTextPane class.

* Added JTalkerMenuBar

Added JTalker Menu Bar. Finally.

* Fixed Codacy issues

Fixed Codacy issues

* Added JTalkerFrame

Added the JFrame to be used by the JTalker application. I am so close to making this look professional and clean. I have a connection to start making (and a UDP server to even setup), but I am so close to the TCP GUI being complete

* Fixed TextPipe bug

I didn't know there was a bug within the text pipe implementation, but I found it and fixed it. Basically, every time you were to add another text updater to the text entry panel, it woul donly work for one of them. But then I realized the issue: the text updater was clearing the text before the others can use it. I could've been lazy and removed that part of the action listener altogether and then only removed it when the enter button was pressed; in the end, I settled for allowing the button to remove the text as its own action (and ensuring this action happened last).

* Finished JTalker GUI

Fixes #6. Finally. This fixes the big gosh darn issue I've been having. It's finally done.